update SWELoss function
Major updates:
1. Make sure loss function work properly (e.g. elevation feed, pred_curve/ref_curve calculation)
2. Use all batch data without basin splitting
3. Update cumulative curve calculation
4. Update _get_elevation and _get_area
5. Add weighted rmse (frequency weight and smooth weight) <- still testing

Base_weight and hypsometry_weight need to be verified. So far base_weight=1.0 and hypsometry_weight=10.0 seem to be the bast combination, while base loss is in the order of 1 and hypsometry loss is in the order of 0.01. Low base_weight may make model learn hypsometry curve (value=0-1) instead of real SWE value (which is not right), leading to underestimation.
Also updates in loss function lead to longer training time while model need more iteration to converge. Epoch=40 may not be sufficient, testing with epoch=100.